fix: move router configuration to separate file
